revisi

+++ app/Http/Controllers/OrderController.php
@@ -4,20 +4,27 @@ namespace App\Http\Controllers;
 
 use App\Models\Order;
 use App\Models\OrderDetail;
-use App\Models\Voucer;
+use Exception;
 use Illuminate\Http\Request;
 
 class OrderController extends Controller
 {
   public function save(Request $request)
   {
-    $model = Order::create($request->all());
-    $this->saveDetail($model->id, $request->items);
+    try {
+      $model = Order::create($request->all());
+      $this->saveDetail($model->id, $request->items);
 
-    return response()->json([
-      'status_code' => 200,
-      'message' => 'Order berhasil dibuat',
-    ]);
+      return response()->json([
+        'status_code' => 200,
+        'message' => 'Order berhasil dibuat',
+      ]);
+    } catch (Exception $e) {
+      return response()->json([
+        'status_code' => 400,
+        'message' => 'Order gagal dibuat',
+      ]);
+    }
   }
 
   public function saveDetail($id, $data)
@@ -34,12 +41,26 @@ class OrderController extends Controller
 
   public function cancel(Request $request, $id)
   {
-    Order::where('id', $id)->delete();
-    OrderDetail::where('order_id', $id)->delete();
+    try {
+      if (Order::where('id', $id)->get()) {
+        Order::where('id', $id)->delete();
+        OrderDetail::where('order_id', $id)->delete();
 
-    return response()->json([
-      'status_code' => 200,
-      'message' => 'Data berhasil dihapus'
-    ]);
+        return response()->json([
+          'status_code' => 200,
+          'message' => 'Data berhasil dihapus'
+        ]);
+      }
+
+      return response()->json([
+        'status_code' => 204,
+        'message' => 'Data tidak ditemukan'
+      ]);
+    } catch (Exception $e) {
+      return response()->json([
+        'status_code' => 400,
+        'message' => 'Order gagal dihapus',
+      ]);
+    }
   }
 }
